+++ Models/Entities/BaseEntity.cs
@@ -19,7 +19,7 @@ public abstract class BaseEntity {
 	[DisplayFormat(DataFormatString = "{0:dd-MM-yyyy HH:mm}", ApplyFormatInEditMode = true)]
 	[Display(Name = "CreatedAt")]
 	public DateTime CreatedAt {
-		get; private set;
+		get; set;
 	} = DateTime.UtcNow;
 
 	/// <summary>