+++ Models/Migrations/20260124162619_InitialCreate.cs
@@ -262,7 +262,7 @@ namespace Models.Migrations
                     OrderId = table.Column<long>(type: "bigint", nullable: false),
                     CoffeeId = table.Column<long>(type: "bigint", nullable: false),
                     Quantity = table.Column<int>(type: "int", nullable: false),
-                    PriceAtPurchase = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
+                    UnitPrice = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
                     CreatedAt = table.Column<DateTime>(type: "datetime2", nullable: false),
                     DeletedAt = table.Column<DateTime>(type: "datetime2", nullable: true)
                 },